Add account setup pages
This commit primarily introduces two new user account setup pages to the application interface, as well as a modal component for account setup. Some visual assets and their accompanying styles are added to back the implementation of these pages. The userSlice - an older model for user data - is deleted, as its functionality is replaced by the aforementioned changes.