Enhance the AccountSetup functionality with Okta authentication
This update allows handling of user access tokens and introduces authentication from the Okta Auth library into the AccountSetup component. In addition, the axios post request in AccountSetup1 has been updated to contain an authorization header for securing the communication. This improvement enhances the security level of the application while setting up an account.