Update Modal component styles and structure
This commit updates the Modal component, adding a new 'contentHeader' CSS class and modifying padding in 'content' class. Also, it introduces changes in the structure of the Modal.tsx file to accommodate the new 'contentHeader' CSS class. The CSS updates enhance the visual design, and file structure changes for the header improve the organization of the code.