Refactor modal to utilize contentContainer
The classification of the "contentContainer" was transferred from "contentHeader" to "content". As a result, the assignment of "overflow-y" and "scrollbar-width" properties in the CSS has now been changed to match this refactoring. The HTML structure and class assignments within the Modal.tsx file have been updated to reflect these changes.