Update certifications structure and fix location fetching
The certifications structure was expanded to include new types and information, but without any message. In the CertificationDetail component, the fetch mechanism was adjusted to consider case difference in certification codes. The functionality for retrieving location information in the LocalResources component was also improved, with more accurate and detailed data now being pulled.